implemented update and delete for userprofile manager #294

+++ server-side/src/main/java/smartcity/accessibility/socialnetwork/UserProfile.java
@@ -55,7 +55,7 @@ public class UserProfile {
 	public double getAvgRating(){
 		if(numOfReviews == 0)
 			return 0;
-		return rating/numOfReviews;
+		return rating/(double)numOfReviews;
 	}
